% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{use_mit_license}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{devtools}{\code{\link[devtools]{use_mit_license}}}
}}

