% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_code_of_conduct}
\alias{use_code_of_conduct}
\title{Add a code of conduct file}
\usage{
use_code_of_conduct(pkg = ".", email = NULL)
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information.}

\item{email}{String. An email address to act as a point of contact for the
project.}
}
\description{
Add the Contributor Covenant version 1.4 code of conduct
}
