% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_ogl_license}
\alias{use_ogl_license}
\title{Add an OGL-3 license}
\usage{
use_ogl_license(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information.}
}
\description{
Adds the necessary infrastructure to declare your package as distributed
under the OGL-3 license.
}
